Code review ammendments
