+++ site/pages/join-us/style.css
@@ -9,28 +9,6 @@
   composes: craft paperTear from "../../components/utils/background.css";
 }
 
-.noWrap {
-  white-space: nowrap;
-}
-
-.headings {
-  padding-top: 24px;
-}
-
-.h1 {
-  composes: serif from "../../css/typography/_fonts.css";
-  composes: fontL from "../../css/typography/_fonts.css";
-  padding-bottom: 10px;
-  text-align: center;
-}
-
-.h3 {
-  composes: fontS2 from '../../css/typography/_fonts.css';
-  composes: serif from '../../css/typography/_fonts.css';
-  margin: 0 0 0.8em;
-  text-align: center;
-}
-
 .bestCompanyLogo {
   margin: 0 auto;
   width: 120px;
@@ -47,14 +25,6 @@
     padding-top: 1em;
   }
   
-  .headPictureContainer {
-    display: none;
-  }
-
-  .headings {
-    padding-top: 0;
-  }
-
   .upperBodyContainer {
     display: flex;
     justify-content: center;
